Rename Singleton meta file
