+++ src/data/singleton_meta.py
